FormConflictMultiple is up and functional.
Removed most functionality because it only
works with Occurrences, not Events. Changing this
to work on an Occurrence-by-Occurrence basis could
result in hundreds of items in the list and be
unmanageable.

So now it only tells you that it conflicts, and doesn't
give you anything to remedy the situation. But that is
enough to allow the user to do it.