*actually* fixed the bug that caused crashes
JSON does not distinguish between integers and longs
but C# parses them as ints until they get too large
longs must be "unboxed" as longs before cast to anything else
same idea with ints.
so when a TimeSpan value is short enough to be an int, it must be
unboxed as an int. Otherwise the program crashes.
shorter values are still ints; longer values go beyond year 9999.

still need to do testing on other computers